working version of kafka communication

+++ ticket_catalogue_service/src/main/kotlin/it/polito/wa2/ticket_catalogue_service/entities/Order.kt
@@ -10,6 +10,6 @@ data class Order(
     var status: Status = Status.PENDING,
     val ticketType: Long, //references Tickets table
     val quantity: Long,
-    val totalPrice: Double,
+    val totalPrice: Float,
     val userId: String
 )
